+++ client/src/App.js
@@ -2,16 +2,15 @@ import { Route, Routes } from "react-router-dom";
 import "bootstrap/dist/css/bootstrap.min.css";
 import React, { Component } from "react";
 import "./App.css";
-import CustomHeader from "./components/customHeader"; //Include Heder
-import Home from "./pages/Home";
-import "./assets/css/content.css";
-import AdminJobs from "./pages/AdminJobs";
-import GradJobs from "./pages/GradJobs";
-import VerifyCertificateNum from "../components/users/VerifyCertificateNum";
-import { Signup } from "./components/Pages/signup/Signup";
+import CustomHeader from "./pages/NavBar/customHeader";
+import Home from "./pages/HomePage/Home";
+import "./pages/HomePage/Slide/content.css";
+import AdminJobs from "./pages/AdminDashboard/AdminJobs";
+import GradJobs from "./pages/GradDashboard/GradJobs";
+import VerifyCertificateNum from "./pages/VerifyCertNum/VerifyCertificateNum";
+import { Signup } from "./pages/SignUpAdmin/Signup";
 import SignUp from "./pages/SignUp/SignUp";
 import JobForm from "./pages/JobForm/JobForm";
-
 class App extends Component {
 	MenuItemTexts = ["STORIES", "INTERVIEW TIPS", "GRADUATE RESOURCES"];
 	render() {
