+++ client/src/pages/AdminDashboard/AdminJobs.js
@@ -1,4 +1,4 @@
-import SearchJobs from "./SearchJobs";
+import SearchJobs from "../JobsList/SearchJobs";
 
 function AdminJobs() {
 	return (
