+++ client/src/pages/HomePage/Slide/cititec-class.jpg
