+++ client/src/pages/HomePage/Slide/content.css
