+++ client/src/pages/HomePage/Slide/slide.js
@@ -1,6 +1,7 @@
 import React from "react";
-import side_image from "../assets/css/cititec-class.jpg";
-import "../assets/css/content.css";
+import side_image from "./cititec-class.jpg";
+import "./content.css";
+
 function Slide() {
 	return (
 		<div className="content">
