+++ client/src/pages/JobsList/SearchJobs.js
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import "./SearchJobs.css";
 import SingleJobListing from "./SingleJobListing";
-import roles from "./roles.json";
+import roles from "../roles.json";
 
 const defaultJobs = roles.jobs;
