+++ client/src/pages/NavBar/SignUpMenu/signupMenu.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "../assets/css/signupMenu.css";
+import "./signupMenu.css";
 
 function SignupMenu() {
 	return (
