+++ client/src/pages/NavBar/cyf_logo.png
