+++ client/src/pages/SignUpAdmin/Buttons/Buttons.js
@@ -1,11 +1,11 @@
 import React from "react";
 import "./Buttons.css";
 
-const Button = ({ onClick, Type = "button", text }) => {
+const Buttons = ({ onClick, Type = "button", text }) => {
 	return (
 		<button type={Type} onClick={onClick}>
 			{text}
 		</button>
 	);
 };
-export default Button;
+export default Buttons;
